fix: update tests for new integral_max formula, add syn_led_next_ms tests
- test_motor_ctrl_saturation: expect 333/-333 (formula-derived integral_max)
  instead of 300/-300 (old heuristic)
- syn_motor_ctrl.c: cast integral_max calc to int64_t to prevent overflow
- test_led.c: add test_led_next_ms covering all modes (OFF, ON, BLINK,
  FLASH, PATTERN with all char types)

+++ src/syntropic/motor/syn_motor_ctrl.c
@@ -169,7 +169,7 @@ SYN_Status syn_motor_ctrl_init(SYN_MotorCtrl *ctrl,
          *   integral_max = output_max * scale * 1000 / ki
          * Use a generous cap to allow full I authority.           */
         .integral_max = (cfg->pid_ki > 0)
-            ? (cfg->output_max * ((int32_t)1 << cfg->pid_scale) * 1000) / cfg->pid_ki
+            ? (int32_t)(((int64_t)cfg->output_max * ((int32_t)1 << cfg->pid_scale) * 1000) / cfg->pid_ki)
             : 0,
         .d_filter_alpha = 200,
     };
